Add skeleton of BoundArgumentsObject
Yesterday it seemed to be a good idea to implement BAO.
However, simpler implementation of SAS were found and
BAO is not needed at the moment (and probably in some
time).

Yet, some work has been done. At least for some time
it may be a good idea to keep the files.

Tests should be OK (apart from the naming convention),
the interface of BAO more or less the same. Only the
implementation is missing, which is usually the easiest.